controller refactoring & navbar, login, signup href update
