+++ src/main/java/com/ll/jigumiyak/admin/AdminController.java
@@ -1,4 +1,18 @@
 package com.ll.jigumiyak.admin;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@RequestMapping("/admin")
+@RequiredArgsConstructor
+@Controller
 public class AdminController {
+
+    @GetMapping("")
+    public String main() {
+        return "administration";
+    }
+
 }
