+++ src/main/java/com/ll/jigumiyak/user/UserController.java
@@ -1,0 +1,24 @@
+package com.ll.jigumiyak.user;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@RequestMapping("/user")
+@RequiredArgsConstructor
+@Controller
+public class UserController {
+
+    private final UserService userService;
+
+    @GetMapping("/login")
+    public String login(){
+        return "login";
+    }
+
+    @GetMapping("/signup")
+    public String signup(){
+        return "signup";
+    }
+}
