+++ src/main/java/com/ll/jigumiyak/user/UserService.java
@@ -1,0 +1,14 @@
+package com.ll.jigumiyak.user;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Service
+public class UserService {
+
+    private final UserRepository userRepository;
+
+
+
+}
