+++ src/main/resources/templates/login.html
@@ -51,7 +51,7 @@
                 <span>카카오 로그인</span>
             </a>
 
-            <a href="#" class="btn btn-block btn-outline border-gray-300 gap-1">
+            <a href="/" class="btn btn-block btn-outline border-gray-300 gap-1">
                 <i class="fa-solid fa-house"></i>
                 <span>메인화면</span>
             </a>
