+++ src/main/resources/templates/navbar.html
@@ -3,7 +3,7 @@
 
     <header class="top-bar">
         <div class="con mx-auto pr-1 h-full flex">
-            <a href="#" class="logo flex items-center px-3">
+            <a href="/" class="logo flex items-center px-3">
               <span>
                 지금이약
               </span>
@@ -18,12 +18,12 @@
                     </a>
                 </li>
                 <li>
-                    <a href="#" class="h-full flex items-center px-2">
+                    <a href="/user/login" class="h-full flex items-center px-2">
                         <span>로그인</span>
                     </a>
                 </li>
                 <li class="menu-1__li-join">
-                    <a href="#" class="h-full flex items-center px-2">
+                    <a href="/user/signup" class="h-full flex items-center px-2">
                         <span>가입하기</span>
                     </a>
                 </li>
