+++ src/main/resources/templates/signup.html
@@ -58,12 +58,12 @@
         <span>카카오 로그인</span>
       </a>
 
-      <a href="#" class="btn btn-block btn-outline border-gray-300 gap-1">
+      <a href="/user/login" class="btn btn-block btn-outline border-gray-300 gap-1">
         <i class="fa-solid fa-right-to-bracket"></i>
         <span>로그인 하러가기</span>
       </a>
 
-      <a href="#" class="btn btn-block btn-outline border-gray-300 gap-1">
+      <a href="/" class="btn btn-block btn-outline border-gray-300 gap-1">
         <i class="fa-solid fa-house"></i>
         <span>메인화면</span>
       </a>
